+++ core/film.cpp
@@ -2,12 +2,11 @@
 // Created by simon on 09.05.17.
 //
 
-#include <cstdint>
 #include "film.hpp"
 
 namespace drdemo {
 
-    Film::Film(uint32_t w, uint32_t h)
+    Film::Film(size_t w, size_t h)
             : width(w), height(h) {}
 
     Film::~Film() {}
